Moved messaging related service class to messaging.service packages
