+++ src/main/kotlin/uk/gov/dluhc/printapi/messaging/ProcessPrintResponseMessageListener.kt
@@ -5,7 +5,7 @@ import mu.KotlinLogging
 import org.springframework.messaging.handler.annotation.Payload
 import org.springframework.stereotype.Component
 import uk.gov.dluhc.printapi.messaging.models.ProcessPrintResponseMessage
-import uk.gov.dluhc.printapi.service.PrintResponseProcessingService
+import uk.gov.dluhc.printapi.messaging.service.PrintResponseProcessingService
 import javax.validation.Valid
 
 private val logger = KotlinLogging.logger { }
