+++ src/main/kotlin/uk/gov/dluhc/printapi/messaging/service/ProcessPrintBatchService.kt
@@ -1,4 +1,4 @@
-package uk.gov.dluhc.printapi.service
+package uk.gov.dluhc.printapi.messaging.service
 
 import org.springframework.stereotype.Service
 import uk.gov.dluhc.printapi.database.entity.Certificate
@@ -6,6 +6,11 @@ import uk.gov.dluhc.printapi.database.entity.PrintRequestStatus.Status.ASSIGNED_
 import uk.gov.dluhc.printapi.database.repository.CertificateRepository
 import uk.gov.dluhc.printapi.database.repository.CertificateRepositoryExtensions.findDistinctByPrintRequestStatusAndBatchId
 import uk.gov.dluhc.printapi.exception.InsufficientPrintRequestsInBatchException
+import uk.gov.dluhc.printapi.service.FilenameFactory
+import uk.gov.dluhc.printapi.service.PrintFileDetailsFactory
+import uk.gov.dluhc.printapi.service.SftpInputStreamProvider
+import uk.gov.dluhc.printapi.service.SftpService
+import uk.gov.dluhc.printapi.service.countPrintRequestsAssignedToBatch
 import javax.transaction.Transactional
 
 /**
